Added Fact model
