+++ slack-nfl-bot/models.rb
@@ -1,1 +1,2 @@
 require 'slack-nfl-bot/models/team'
+require 'slack-nfl-bot/models/fact'
